Moving reporting backend specific code to specific packages
